+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 - The next release will be 1.0.0, which means that master will contain breaking changes.
 
 *Enhancements*
-- `krane deploy` and `krane render` allow deploy filenames given from stdin stream with given `--stdin` argument. ([#630](https://github.com/Shopify/kubernetes-deploy/pull/630))
+- `krane deploy` and `krane render` allow template filenames given from stdin stream with given `--stdin` argument. ([#630](https://github.com/Shopify/kubernetes-deploy/pull/630))
 
 ## 0.31.1